+++ rxjava2/src/main/java/io/vertx/reactivex/core/RxHelper.java
@@ -3,8 +3,8 @@ package io.vertx.reactivex.core;
 import io.reactivex.Flowable;
 import io.reactivex.Observable;
 import io.vertx.core.streams.ReadStream;
-import io.vertx.reactivex.FlowableReadStream;
-import io.vertx.reactivex.ObservableReadStream;
+import io.vertx.reactivex.core.impl.FlowableReadStream;
+import io.vertx.reactivex.core.impl.ObservableReadStream;
 
 import java.util.function.Function;
