+++ rxjava2/src/main/java/io/vertx/reactivex/core/impl/FlowableReadStream.java
@@ -1,4 +1,4 @@
-package io.vertx.reactivex;
+package io.vertx.reactivex.core.impl;
 
 import io.reactivex.Flowable;
 import io.reactivex.FlowableSubscriber;
